対話式のシェルを実行する関数を interactive_shell() と名付けた
