+++ minishell.h
@@ -28,7 +28,6 @@ typedef struct s_cmd_str_node {
 // FreeBSD doesn't have type __sighandler_t.
 typedef void	(*t_sighandler)(int);
 
-// AST to command_invocation
 t_command_invocation	*cmd_ast_pipcmds2cmdinvo(t_parse_node_pipcmds *pipcmds);
 t_command_invocation	*cmd_ast_cmd2cmdinvo(t_parse_node_command *cmd_node);
 char					*string_node2string(t_parse_node_string *string_node);
@@ -45,6 +44,7 @@ int						put_minish_err_msg_and_ret(int ret_val,
 void					put_minish_err_msg_and_exit(int status,
 							const char *cmd_name, const char *msg);
 int						invoke_sequential_commands(t_parse_ast *seqcmd);
+int						interactive_shell(void);
 
 // Shell initialization
 void					init_g_shell(void);
